+++ STranslate.Plugin.Translate.DeepSeek/Main.cs
@@ -132,7 +132,7 @@ public class Main : LlmTranslatePluginBase
 
         // 选择模型
         var model = Settings.Model.Trim();
-        model = string.IsNullOrEmpty(model) ? "deepseek-chat" : model;
+        model = string.IsNullOrEmpty(model) ? "deepseek-v4-flash" : model;
 
         // 替换Prompt关键字
         var messages = (Prompts.FirstOrDefault(x => x.IsEnabled) ?? throw new Exception("请先完善Propmpt配置"))