+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 更新
 
-- 更换版本tag #1
+- 模型下拉菜单名称更改为的"deepseek-v4-flash"与"deepseek-v4-pro"